+++ packages/atomic/src/components/atomic-rating/atomic-rating.tsx
@@ -17,7 +17,7 @@ export const Rating: FunctionalComponent<RatingProps> = (props) => {
     return (
       <atomic-icon
         icon={props.icon}
-        class={active ? 'icon-active' : 'icon-inactive'}
+        class={`flex-shrink-0 ${active ? 'icon-active' : 'icon-inactive'}`}
         style={{width: iconSize, height: iconSize}}
       ></atomic-icon>
     );