+++ packages/quantic/force-app/main/default/lwc/quanticResultsPerPage/quanticResultsPerPage.js
@@ -40,12 +40,8 @@ export default class QuanticResultsPerPage extends LightningElement {
   }
 
   disconnectedCallback() {
-    if (this.unsubscribe) {
-      this.unsubscribe();
-    }
-    if (this.unsubscribeSearchStatus) {
-      this.unsubscribeSearchStatus();
-    }
+    this.unsubscribe?.();
+    this.unsubscribeSearchStatus?.();
   }
 
   updateState() {